240321 1252 생산 이벤트리스너 완성

+++ sixbeam/src/main/java/com/erpproject/sixbeam/pd/dto/OrderDto.java
@@ -20,7 +20,7 @@ OrderDto {
     private LocalDate orderDelivDt;
     private EmpInfoEntity empInfoEntity;
     private ItemEntity itemEntity;
-    private Long orderAmt;
+    private int orderAmt;
     private boolean orderSt;
 
     public OrderEntity toEntity() {
